interim commit; debugging startup
Added a few tips/hints to Readme.md for directing the necessary
modifications to /etc/nixos/configuration.nix. Also converted
back to the use of the UNCLE which has been fixed to perform an
overlay mount of /nix/store, thus no longer preventing access
to configuration files linked from /etc/flox.toml into the store.
This has in turn re-enabled the "init" daemon functionality of
floxd, which (happily) brings the startup on NixOS in line with
all other O/S variants.

We had an issue whereby the `unshare` invocation failed to perform
the overlay mount, complaining that /flox/store wasn't ready, but
that disappeared when we added "set -x" to the unshare script
invocation. Expect for that to return at some point ...