Make resilient to nixos-rebuild, systemctl (stop|restart)
We were having the problem that nixox-rebuild invocations were killing ICFS
and thus breaking any open filehandles held by running processes.  This
update fixes that problem by ensuring that ICFS is only ever restarted by
way of the `floxadm restart` command, in the two scenarios:

 1. when running `nixos-rebuild`, invoke `systemctl reload` to perform the
    `reload` action, which is configured to run `floxadm restart`
 2. when running `systemctl (stop|restart)`, configure the `ExecStop` action
    to run `floxadm restart` and `KillMode = "none"` to prevent systemd
    from killing any of the flox processes directly

In future we will update ICFS to use the FileDescriptorStore systemd feature
that allows it to park an open filehandle for /dev/fuse so that subsequent
restarts can pick up the open filehandles and processes won't crash. Once we
do that we can keep `reloadIfChanged = true` but we won't need to define an
`ExecStop` action because ICFS will "do the right thing" upon receipt of
SIGTERM.